feat: implement module subtask
